chore: consolidate test utils
