+++ test/integration/simple_zonal_with_asm/simple_zonal_with_asm_test.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/gcloud"
 	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/tft"
+	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/utils"
 	"github.com/gruntwork-io/terratest/modules/k8s"
 	"github.com/stretchr/testify/assert"
 	"github.com/terraform-google-modules/terraform-google-kubernetes-engine/test/integration/testutils"
@@ -49,11 +50,11 @@ func TestSimpleZonalWithASM(t *testing.T) {
 		k8sOpts := k8s.KubectlOptions{}
 		listNameSpace, err := k8s.RunKubectlAndGetOutputE(t, &k8sOpts, "get", "ns", "istio-system", "-o", "json")
 		assert.NoError(err)
-		kubeNS := testutils.ParseKubectlJSONResult(t, listNameSpace)
+		kubeNS := utils.ParseKubectlJSONResult(t, listNameSpace)
 		assert.Contains(kubeNS.Get("metadata.name").String(), "istio-system", "Namespace is Functional")
 		listConfigMap, err := k8s.RunKubectlAndGetOutputE(t, &k8sOpts, "get", "configmap", "asm-options", "-n", "istio-system", "-o", "json")
 		assert.NoError(err)
-		kubeCM := testutils.ParseKubectlJSONResult(t, listConfigMap)
+		kubeCM := utils.ParseKubectlJSONResult(t, listConfigMap)
 		assert.Contains(kubeCM.Get("metadata.name").String(), "asm-options", "Configmap is Present")
 
 	})
